Move serial workflow to controller
